Add tags styling and remove file from VCS
This commit introduces styling for tags in CSS, adds tag based structure in list.html, tags.html and terms.html files. A command has been modified in package.json. This was done to add visual enhancements to the tags used in the web pages and to improve the overall user experience. Moreover, tailwind compiled css file has been added to .gitignore to prevent unwanted file to be versioned.